+++ client/src/components/Checkout/Checkout.jsx
@@ -13,8 +13,9 @@ function Checkout({ products, data, total }) {
     script.setAttributeNode(attr_data_preference);
 
     document.getElementById("form1").appendChild(script);
-    // return () => {
-    // };
+     return () => {
+      //  document.getElementById("form1").removeChild(script);
+     };
   }, [data]);
 
   return (
