+++ client/src/redux/combineReducers.js
@@ -7,10 +7,10 @@ import brandReducer from './brand/brandReducer';
 import categoriesReducer from './category/categoriesReducer';
 import sizeReducer from './sizes/sizeReducer';
 import newsletterReducer from './newsletter/newsletterReducer';
-import cartReducersTomi from './cartTomi/cartReducerTomi';
+import cartReducers from './cart/cartReducer';
 import ratingReducer from './rating/ratingReducer';
 import ordersReducer from './orders/ordersReducer';
-import cartReducer from './cart/cartReducer';
+
 
 export const reducers = combineReducers({
     reviewsReducer,
@@ -20,8 +20,7 @@ export const reducers = combineReducers({
     categoriesReducer,
     sizeReducer,
     newsletterReducer,
-    cartReducer,
-    cartReducersTomi,
+    cartReducers,
     ratingReducer,
     ordersReducer
 });
